requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
Django==4.1.7
django-cache==0.3
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-formtools==2.4
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-phonenumber-field==7.0.2
django-tinymce==3.5.0
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
phonenumberslite==8.13.7
Pillow==9.4.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
